feat: implement command interaction cooldown (rate limit) (#12)
* feat(Command): add cooldown property

Add a 'cooldown' property to the Command class. This property represents the cooldown period for the command in milliseconds, determining the minimum amount of time that must pass before the command can be executed again. The default value is set to 2500 milliseconds if not provided in the command options.

* feat: implement command cooldowns

Added command cooldown functionality. This includes a new 'cooldowns' collection to track user cooldowns, a 'cooldownSweep' interval to clear expired cooldowns, and a 'userInCooldown' method to handle interactions from users currently in cooldown.